provide connectivity for ubuntu/kali/debian OS-es
TODO: The current Ubuntu vm in the sample test suite has difficulty
finishing the unattended install even though it used to do so in
the past. The fact that it is used as a permanent vm in the sample
test suite already makes this lower priority issue but it could
still be combined with a proper Ubuntu version update.