todos from dropped self-sync complexity ("own" pool scope filtering)
It turns out that self-sync will only filter out some cases where
"own" pool scope is supposed to be dropped as it can also be used
when syncing between two regular distinct test nodes. The disbanded
parameter could enter the pool consideration even when node A syncs
from node B and self-sync condition is not triggered. I don't know
where my previous thinking was in here before.

NOTES:
------

TODO: The problem we encounter at the moment is a chain of mutually
hiding bugs - a state show that includes shared pool sources without
the states will intersect them with the ones from a previous worker
and end up finding nothing, resulting in TestAbort for the sync-get.
Since the swarms are currently implicitly using a swarm pool this is
not a problem and it wasn't a problem before for the same reason. In
fact the above error was also hidden before since we would not filter
out "own" pool scope thus letting the state show unionizing with the
existing cache states and not raising the error. An actual copy of
the states would fail in both the before case and current case though.

TODO: The solution would rather be to let the state show simply give
all information without AND/OR assumptions and let each operation
like get/set/unset decide how to use this information on their own
and thus drop the check() and state policies from state_setup. Or
perhaps less fatal, refine the state_setup check()'s use of the
state show() so that we pool-AND + cache-OR just there for the
decision making of the state policies.

TODO: Make shared swarm location obligatory? Or just make self-sync
native.