Rely on instructive show_states instead of binary restrictive check_states
Drop the binary single-state check performing check_states function
for a more versatile show_states function that can include the root
pseudo-state as one of the shown states and provide more information
at the same time via a filter regex matching more than one state name.
This in turns requires strict use of the root keyword as it is now
returned as a shown state.

The previous root state handling will now be handled in show_states
before it is made more specific to selected state backends.

Some order of the shown states could not be guaranteed and needed
explicit set-list conversions.