Move image processing from early to late vm processing hooks
As image state backends now can switch off and on a vm they need
the vm's environment setup to be the same as the one for vm state
backends that switch on and off vms even before this.

This introduces an odd issue where tests with ramfile loaded vms
will pass and fail in a cycle because the vm loaded from ramfile
will end up with incompatible CPU. It turns out that the state get
params are correct and consistent but the vm.params end up containing
a correct EPYC auto-detected CPU model followed by a wrong SandyBridge
one. These get loaded from the vm environment (env) and are correctly
set in any vm created by the env_process but the VMNetwork updates
the CPU model to a wrong SandyBridge as its parameters are not yet
ready and don't reflect the complete preprocessing (net states are
retrieved during the early preprocessing hook). They still get saved
in the env and reloaded with wrong values leading to crash when a vm
state with a different CPU model is loaded from ramfile.

The problem was likely fixed before by the early preprocessing hook
image state ops as these killed the vms which were then recreated by
env_process at the right time (when ready) and moving the image state
ops to a later stage means that the vm with unprepared params is left
running and thus not updated by the env_process with CPU model and
other preprocessing.

We get passing followed by failing tests as on each fail the vms are
recreated thus killed with follow up run getting good parameter vms
which are then kept running in the env to be overwritten with bad
parameters afterwards.

TODO: The circular vmnet dependence on vm objects that in turn depend
on proper bridges and interfaces can be at least partly remedied by
moving the net state ops to the late preprocessing hook.