Drop vm state root from LVM state root
LVM set_root also creates a qcow2 image which should instead be
separately checked and created by qcow2vt it could be combined
with. The new show policy will always set such a QCOW2 image as
precondition needed for the vm states when a vm state is created
and these preconditions are not met thus decoupling from the LVM
image backend if necessary.