Fix cleaning the same guisetup states due to multi-vm1 variants
The following makes set_state_images_vm2 safely separated for each
vm1 variant by moving both the default and non-default cases in
separate conditional blocks.

Make sure that the GUI test script now works based on test name
and not on vm2 set state which might not always be defined.

Work around a limited conditional parameter overwriting by the
Cartesian parser. It has the old limitation where a parameter that
was unconditionally defined cannot be overwritten in any way in a
conditional block.