Recover better default get policy we could not implement so far
Now that we have qcow2ext-specific pointer head image handling we
don't have to provide it as part of the preconditions using force
get policy and can finally allow the initially desired default get
policy.

TODO: This at present works for vm3 as a permanently shared vm but
will not work for any qcow2/qcow2vt based non-permanent vm as the
state sync seems broken. The latter would become obvious as soon
as we extend our coverage to test for it.