update integration test path comparison to swarm pool
The local path sync is not tested for a long time now.

Note that we cannot diff compare only paths of possibly different
files so we need to use `find` to list the files in the directories.

TODO: This finds an actual error where qcow2/qcow2vt do not sync
states across workers properly.